put it all together
