--- data.js
@@ -1,266 +1,0 @@
-const data = [
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: 'ME',
-    sku: '',
-    sub_total: 10.0,
-    buyer_note: null,
-    package_id: 1460351373,
-    ship_date: '06 Sep 2016 20:51:21 GMT',
-    ship_to_zip: '04103',
-    payment_id: 2725873591,
-    tax: null,
-    ship_to_name: 'Matthew Glidden',
-    paypal_id: '58U341853S160894V',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: 'matthew.glidden311@gmail.com',
-    order_date: '02 Sep 2016 16:56:31 GMT',
-    order_total: 10.0,
-    shipping: null,
-    ship_to_country: 'United States',
-    ship_to_street: '99 Euclid Ave',
-    buyer_name: 'Matthew Glidden',
-    item_url: 'https://crucialrip.bandcamp.com/merch/crucial-rip-logo-t-shirt',
-    item_name: 'Crucial Rip Logo T-shirt by Crucial Rip',
-    option: 'Extra-Large',
-    ship_to_street_2: null,
-    sale_item_id: 37564101,
-    option_id: 209321616,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Portland',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 5,
-    ship_to_state: 'ME',
-    sku: '',
-    sub_total: 10.0,
-    buyer_note: null,
-    package_id: 799316743,
-    ship_date: '06 Sep 2016 20:51:21 GMT',
-    ship_to_zip: '04103',
-    payment_id: 2725873591,
-    tax: null,
-    ship_to_name: 'Matthew Glidden',
-    paypal_id: '58U341853S160894V',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: 'matthew.glidden311@gmail.com',
-    order_date: '02 Sep 2016 16:56:31 GMT',
-    order_total: 10.0,
-    shipping: null,
-    ship_to_country: 'United States',
-    ship_to_street: '99 Euclid Ave',
-    buyer_name: 'Matthew Glidden',
-    item_url:
-      'https://crucialrip.bandcamp.com/merch/crucial-rip-gorebong-sticker',
-    item_name: 'Crucial Rip Gorebong Sticker by Crucial Rip',
-    option: null,
-    ship_to_street_2: null,
-    sale_item_id: 37564139,
-    option_id: null,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Portland',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: 'ME',
-    sku: '',
-    sub_total: 7.0,
-    buyer_note: null,
-    package_id: 507574767,
-    ship_date: '06 Sep 2016 20:51:21 GMT',
-    ship_to_zip: '04103',
-    payment_id: 2725873591,
-    tax: null,
-    ship_to_name: 'Matthew Glidden',
-    paypal_id: '58U341853S160894V',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: 'matthew.glidden311@gmail.com',
-    order_date: '02 Sep 2016 16:56:31 GMT',
-    order_total: 7.0,
-    shipping: null,
-    ship_to_country: 'United States',
-    ship_to_street: '99 Euclid Ave',
-    buyer_name: 'Matthew Glidden',
-    item_url: 'https://crucialrip.bandcamp.com/album/bong-rip-evisceration',
-    item_name: 'Bong Rip Evisceration: Bong Rip Evisceration CD by Crucial Rip',
-    option: null,
-    ship_to_street_2: null,
-    sale_item_id: 37564156,
-    option_id: null,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Portland',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: null,
-    sku: '',
-    sub_total: 7.0,
-    buyer_note: null,
-    package_id: 507574767,
-    ship_date: '07 Mar 2017 22:09:37 GMT',
-    ship_to_zip: '8009',
-    payment_id: 2419592830,
-    tax: null,
-    ship_to_name: 'Erik Helgesen',
-    paypal_id: '5FX32492R9019051T',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'NO',
-    buyer_email: 'Erikhelgesen2000@gmail.com',
-    order_date: '28 Feb 2017 17:23:02 GMT',
-    order_total: 12.0,
-    shipping: 5.0,
-    ship_to_country: 'Norway',
-    ship_to_street: 'Leitetunet 9 H303',
-    buyer_name: 'Erik Helgesen',
-    item_url: 'https://crucialrip.bandcamp.com/album/bong-rip-evisceration',
-    item_name: 'Bong Rip Evisceration: Bong Rip Evisceration CD by Crucial Rip',
-    option: null,
-    ship_to_street_2: null,
-    sale_item_id: 45163853,
-    option_id: null,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Bodø',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: 'TX',
-    sku: '',
-    sub_total: 15.0,
-    buyer_note: null,
-    package_id: 1648928814,
-    ship_date: '07 Feb 2018 20:56:59 GMT',
-    ship_to_zip: '78745',
-    payment_id: 963215293,
-    tax: null,
-    ship_to_name: 'Alexander Ross',
-    paypal_id: '6J7314204D6695124',
-    ship_notes: 'These are some shipping notes',
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: '30tco96@gmail.com',
-    order_date: '02 Feb 2018 03:23:41 GMT',
-    order_total: 20.0,
-    shipping: 5.0,
-    ship_to_country: 'United States',
-    ship_to_street: '412 E William Cannon',
-    buyer_name: 'Alexander Ross',
-    item_url: 'https://crucialrip.bandcamp.com/merch/logo-tee-with-vadm-back',
-    item_name: 'Logo Tee with VADM Back by Crucial Rip',
-    option: 'Green - Large',
-    ship_to_street_2: 'Apt. 805',
-    sale_item_id: 60397716,
-    option_id: 1583729963,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Austin',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: 'TX',
-    sku: '',
-    sub_total: 20.0,
-    buyer_note: null,
-    package_id: 3934011684,
-    ship_date: '07 Feb 2018 20:56:12 GMT',
-    ship_to_zip: '78745',
-    payment_id: 1942243563,
-    tax: null,
-    ship_to_name: 'Alexander Ross',
-    paypal_id: '1VS43761AN1250323',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: '30tco96@gmail.com',
-    order_date: '02 Feb 2018 03:25:46 GMT',
-    order_total: 25.0,
-    shipping: 5.0,
-    ship_to_country: 'United States',
-    ship_to_street: '412 E William Cannon',
-    buyer_name: 'Alexander Ross',
-    item_url:
-      'https://crucialrip.bandcamp.com/merch/hash-slinging-slasher-t-shirt',
-    item_name: 'Hash Slinging Slasher T-shirt by Crucial Rip',
-    option: 'Large',
-    ship_to_street_2: 'Apt. 805',
-    sale_item_id: 60397781,
-    option_id: 452508172,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Austin',
-    discount_code: null,
-    buyer_phone: null,
-  },
-  {
-    payment_state: 'paid',
-    quantity: 1,
-    ship_to_state: 'OR',
-    sku: '',
-    sub_total: 10.0,
-    buyer_note: null,
-    package_id: 3934011684,
-    ship_date: '17 Mar 2018 14:00:02 GMT',
-    ship_to_zip: '97306',
-    payment_id: 469685921,
-    tax: null,
-    ship_to_name: 'Joe Blacksmith',
-    paypal_id: '9XD21389SR519840B',
-    ship_notes: null,
-    currency: 'USD',
-    ship_to_country_code: 'US',
-    buyer_email: 'choptop645@hotmail.com',
-    order_date: '13 Mar 2018 18:57:24 GMT',
-    order_total: 15.0,
-    shipping: 5.0,
-    ship_to_country: 'United States',
-    ship_to_street: '2000 Robins Ln. SE. #72',
-    buyer_name: 'Joseph Blacksmith',
-    item_url:
-      'https://crucialrip.bandcamp.com/merch/hash-slinging-slasher-t-shirt',
-    item_name: 'Hash Slinging Slasher T-shirt by Crucial Rip',
-    option: 'Xtra Large',
-    ship_to_street_2: null,
-    sale_item_id: 62503775,
-    option_id: null,
-    selling_band_url: 'https://crucialrip.bandcamp.com',
-    ship_from_country_name: 'United States',
-    artist: 'Crucial Rip',
-    ship_to_city: 'Salem',
-    discount_code: null,
-    buyer_phone: null,
-  },
-];
-
-module.exports = data;
