+++ src/index.js
@@ -1,18 +1,22 @@
 require('dotenv').config();
-const { postSlackMessage } = require('./slack');
 const BandcampClient = require('./bandcamp');
 
 let client = new BandcampClient();
 
-const main = async (client) => {
-  console.log(await client.getOrders(process.env.BANDCAMP_BAND_ID));
+const checkForOrders = async (client) => {
+  let orders = await client.getOrders(process.env.BANDCAMP_BAND_ID);
+  if (orders?.success === true) {
+    if (orders.items.length === 0) return;
+    orderDetected(orders.items);
+  } else {
+    console.log('Something went wrong');
+  }
 };
 
-let minutes = 0.25;
-
+let minutes = 0.5;
 let interval = minutes * 60 * 1000;
 
 setInterval(() => {
   console.log('Checking...');
-  main(client);
+  checkForOrders(client);
 }, interval);
